requirements.txt
groq>=0.9.0
apscheduler>=3.10.4
requests>=2.31.0
python-dotenv>=1.0.0
